Update and rename test_mmap.c to test_NNQAK.c
-----BEGIN PGP SIGNATURE-----

iQFFBAEBCgAvFiEEbsyLEzg/qUTA43uogItw+93Q3WUFAmdMmQwRHG1jckBzYW5k
ZWxtYW4uY2EACgkQgItw+93Q3WWu3QgArDI5uqSaFfwJ41K+zw2dM/+5w9y2VafS
+sULzuTe6pMA+DJqeTC9TcdPl3zOaORwApw545q9nezUmUG0QOCYO7npRiKOYRH2
4jZZ4vbuvS1fhFAicAvWSdGJRL95JGkpqFbklHLS53xoouXDE4Ze8Fp6r4sgl4LN
onNIrDnmEB21RSZ2b5V2YhDdLsivfnk17+XSp8dhp9xBtPa0VVhDLItr2wZeSb9H
3BESMzHFvPRHI0k6wasBp0ymGFCzAbukqe0PoM9mHfU6F06rrtQjW/HwosLkkGcR
iT/SnMYitcT0bgLZ1skzaola+Epx4TLlapkxjvZlkCQVQr9DDISeVQ==
=3Nk/
-----END PGP SIGNATURE-----
